Chapter 09/requirements.txt
streamlit
chromadb
llama-index
langchain
langchain-community
langchain-nomic
sentence-transformers
transformers
pillow
pypdf
scikit-learn
numpy